ignition: Split fragment types into a dedicated module
Move the Ignition config structs and serialization into src/ignition.rs
with a write_to() method on IgnitionConfig, leaving the
provider-to-fragment orchestration in the render-ignition CLI.
