+++ src/cli/render_ignition.rs
@@ -1,24 +1,17 @@
 //! `render-ignition` CLI sub-command.
 //!
-//! Generates per-feature Ignition config fragment files in an output directory.
-//! Each enabled feature writes its own `.ign` file; Ignition merges them
-//! natively from `base.platform.d/<platform>/` under a system config directory
-//! such as `/etc/ignition`.
-//!
-//! The Ignition config fragment types and `write_fragment`/`hostname_data_uri`
-//! helpers are provider-agnostic and live here so other platforms can reuse
-//! them; any provider-specific data (e.g. the admin password) is sourced
-//! through the `MetadataProvider` trait rather than referenced directly.
+//! Fetches metadata from a cloud provider and writes Ignition config fragment
+//! files for the enabled features. The fragment types and their serialization
+//! live in [`crate::ignition`]; this module turns provider metadata into those
+//! fragments and handles argument parsing and dispatch.
 
 use anyhow::{bail, Context, Result};
 use clap::{ArgGroup, Parser};
-use serde::Serialize;
 use slog_scope::{info, warn};
-use std::fs;
-use std::os::unix::fs::PermissionsExt;
 use std::path::Path;
 
-const IGNITION_VERSION: &str = "3.0.0";
+use crate::ignition::IgnitionConfig;
+use crate::providers::MetadataProvider;
 
 /// Render Ignition config fragments from cloud provider metadata
 #[derive(Debug, Parser)]
@@ -80,78 +73,7 @@ impl CliRenderIgnition {
     }
 }
 
-#[derive(Debug, Serialize)]
-pub(crate) struct IgnitionConfig {
-    pub ignition: IgnitionMeta,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub storage: Option<Storage>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub passwd: Option<Passwd>,
-}
-
-#[derive(Debug, Serialize)]
-pub(crate) struct IgnitionMeta {
-    pub version: String,
-}
-
-#[derive(Debug, Serialize)]
-pub(crate) struct Storage {
-    pub files: Vec<StorageFile>,
-}
-
-#[derive(Debug, Serialize)]
-pub(crate) struct StorageFile {
-    pub path: String,
-    pub mode: u32,
-    pub overwrite: bool,
-    pub contents: FileContents,
-}
-
-#[derive(Debug, Serialize)]
-pub(crate) struct FileContents {
-    pub source: String,
-}
-
-#[derive(Debug, Serialize)]
-pub(crate) struct Passwd {
-    pub users: Vec<PasswdUser>,
-}
-
-#[derive(Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub(crate) struct PasswdUser {
-    name: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub ssh_authorized_keys: Option<Vec<String>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub password_hash: Option<String>,
-}
-
-pub(crate) fn hostname_data_uri(hostname: &str) -> String {
-    let encoded =
-        percent_encoding::utf8_percent_encode(hostname, percent_encoding::NON_ALPHANUMERIC)
-            .to_string();
-    format!("data:,{encoded}")
-}
-
-pub(crate) fn write_fragment(config: &IgnitionConfig, path: &Path) -> Result<()> {
-    if let Some(parent) = path.parent() {
-        fs::create_dir_all(parent)
-            .with_context(|| format!("failed to create directory {}", parent.display()))?;
-    }
-    let json =
-        serde_json::to_string_pretty(config).context("failed to serialize ignition config")?;
-    fs::write(path, json.as_bytes())
-        .with_context(|| format!("failed to write {}", path.display()))?;
-    fs::set_permissions(path, fs::Permissions::from_mode(0o644))
-        .with_context(|| format!("failed to set permissions on {}", path.display()))?;
-    Ok(())
-}
-
-fn generate_hostname_fragment(
-    provider: &dyn crate::providers::MetadataProvider,
-    output_dir: &str,
-) -> Result<()> {
+fn generate_hostname_fragment(provider: &dyn MetadataProvider, output_dir: &str) -> Result<()> {
     let hostname = match provider.hostname()? {
         Some(h) => h,
         None => {
@@ -160,33 +82,13 @@ fn generate_hostname_fragment(
         }
     };
 
-    let config = IgnitionConfig {
-        ignition: IgnitionMeta {
-            version: IGNITION_VERSION.to_string(),
-        },
-        storage: Some(Storage {
-            files: vec![StorageFile {
-                path: "/etc/hostname".into(),
-                mode: 420,
-                overwrite: true,
-                contents: FileContents {
-                    source: hostname_data_uri(&hostname),
-                },
-            }],
-        }),
-        passwd: None,
-    };
-
     let path = Path::new(output_dir).join("hostname.ign");
-    write_fragment(&config, &path)?;
+    IgnitionConfig::hostname_fragment(&hostname).write_to(&path)?;
     info!("wrote hostname ignition fragment"; "path" => path.display().to_string());
     Ok(())
 }
 
-fn generate_user_fragment(
-    provider: &dyn crate::providers::MetadataProvider,
-    output_dir: &str,
-) -> Result<()> {
+fn generate_user_fragment(provider: &dyn MetadataProvider, output_dir: &str) -> Result<()> {
     let username = provider
         .admin_username()
         .context("failed to query admin username from provider")?;
@@ -209,26 +111,8 @@ fn generate_user_fragment(
         .admin_password_hash()
         .context("failed to query admin password hash from provider")?;
 
-    let config = IgnitionConfig {
-        ignition: IgnitionMeta {
-            version: IGNITION_VERSION.to_string(),
-        },
-        storage: None,
-        passwd: Some(Passwd {
-            users: vec![PasswdUser {
-                name: username,
-                ssh_authorized_keys: if ssh_keys.is_empty() {
-                    None
-                } else {
-                    Some(ssh_keys)
-                },
-                password_hash,
-            }],
-        }),
-    };
-
     let path = Path::new(output_dir).join("user.ign");
-    write_fragment(&config, &path)?;
+    IgnitionConfig::user_fragment(username, ssh_keys, password_hash).write_to(&path)?;
     info!("wrote platform-user ignition fragment"; "path" => path.display().to_string());
     Ok(())
 }
@@ -237,6 +121,7 @@ fn generate_user_fragment(
 mod tests {
     use super::*;
     use openssh_keys::PublicKey;
+    use std::fs;
 
     /// Minimal generic provider used to exercise the fragment generators
     /// without any platform-specific plumbing.
@@ -247,7 +132,7 @@ mod tests {
         admin_password_hash: Option<String>,
     }
 
-    impl crate::providers::MetadataProvider for FakeProvider {
+    impl MetadataProvider for FakeProvider {
         fn hostname(&self) -> Result<Option<String>> {
             Ok(self.hostname.clone())
         }
@@ -350,146 +235,5 @@ mod tests {
 
         generate_user_fragment(&provider, dir).unwrap();
         assert!(!tmp.path().join("user.ign").exists());
-    }
-
-    #[test]
-    fn test_ignition_json_with_keys() {
-        let cfg = IgnitionConfig {
-            ignition: IgnitionMeta {
-                version: "3.0.0".into(),
-            },
-            storage: None,
-            passwd: Some(Passwd {
-                users: vec![PasswdUser {
-                    name: "testuser".into(),
-                    ssh_authorized_keys: Some(vec!["ssh-ed25519 AAAA...".into()]),
-                    password_hash: None,
-                }],
-            }),
-        };
-        let v: serde_json::Value =
-            serde_json::from_str(&serde_json::to_string_pretty(&cfg).unwrap()).unwrap();
-        assert_eq!(v["ignition"]["version"], "3.0.0");
-        assert_eq!(v["passwd"]["users"][0]["name"], "testuser");
-        assert_eq!(
-            v["passwd"]["users"][0]["sshAuthorizedKeys"][0],
-            "ssh-ed25519 AAAA..."
-        );
-        assert!(v["passwd"]["users"][0].get("passwordHash").is_none());
-    }
-
-    #[test]
-    fn test_ignition_json_with_password_hash() {
-        let cfg = IgnitionConfig {
-            ignition: IgnitionMeta {
-                version: "3.0.0".into(),
-            },
-            storage: None,
-            passwd: Some(Passwd {
-                users: vec![PasswdUser {
-                    name: "azureuser".into(),
-                    ssh_authorized_keys: None,
-                    password_hash: Some("$6$rounds=10000$salt$hash".into()),
-                }],
-            }),
-        };
-        let v: serde_json::Value =
-            serde_json::from_str(&serde_json::to_string_pretty(&cfg).unwrap()).unwrap();
-        assert_eq!(v["passwd"]["users"][0]["name"], "azureuser");
-        assert!(v["passwd"]["users"][0].get("sshAuthorizedKeys").is_none());
-        assert_eq!(
-            v["passwd"]["users"][0]["passwordHash"],
-            "$6$rounds=10000$salt$hash"
-        );
-    }
-
-    #[test]
-    fn test_write_fragment_emits_valid_json_and_permissions() {
-        let tmp = tempfile::tempdir().unwrap();
-        let out_file = tmp
-            .path()
-            .join("etc/ignition/base.platform.d/azure/extensions.ign");
-
-        let cfg = IgnitionConfig {
-            ignition: IgnitionMeta {
-                version: "3.0.0".into(),
-            },
-            storage: None,
-            passwd: Some(Passwd {
-                users: vec![PasswdUser {
-                    name: "core".into(),
-                    ssh_authorized_keys: Some(vec![
-                        "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAAAgQDYVEprvtYJXVOBN0XNKVVRNCRX6BlnNbI+USLGais1sUWPwtSg7z9K9vhbYAPUZcq8c/s5S9dg5vTHbsiyPCIDOKyeHba4MUJq8Oh5b2i71/3BISpyxTBH/uZDHdslW2a+SrPDCeuMMoss9NFhBdKtDkdG9zyi0ibmCP6yMdEX8Q== Generated by Nova".into(),
-                    ]),
-                    password_hash: None,
-                }],
-            }),
-        };
-
-        write_fragment(&cfg, &out_file).unwrap();
-
-        assert!(out_file.exists());
-
-        let raw = fs::read_to_string(&out_file).unwrap();
-        let json: serde_json::Value = serde_json::from_str(&raw).unwrap();
-        assert_eq!(json["ignition"]["version"], "3.0.0");
-        assert_eq!(json["passwd"]["users"][0]["name"], "core");
-
-        let mode = fs::metadata(&out_file).unwrap().permissions().mode() & 0o777;
-        assert_eq!(mode, 0o644);
-    }
-
-    #[test]
-    fn test_hostname_data_uri() {
-        assert_eq!(hostname_data_uri("core1"), "data:,core1");
-        assert_eq!(
-            hostname_data_uri("my-vm.internal"),
-            "data:,my%2Dvm%2Einternal"
-        );
-    }
-
-    #[test]
-    fn test_hostname_storage_fragment_serialization() {
-        let cfg = IgnitionConfig {
-            ignition: IgnitionMeta {
-                version: "3.0.0".into(),
-            },
-            storage: Some(Storage {
-                files: vec![StorageFile {
-                    path: "/etc/hostname".into(),
-                    mode: 420,
-                    overwrite: true,
-                    contents: FileContents {
-                        source: hostname_data_uri("myvm"),
-                    },
-                }],
-            }),
-            passwd: None,
-        };
-
-        let v: serde_json::Value =
-            serde_json::from_str(&serde_json::to_string_pretty(&cfg).unwrap()).unwrap();
-        assert_eq!(v["ignition"]["version"], "3.0.0");
-        assert_eq!(v["storage"]["files"][0]["path"], "/etc/hostname");
-        assert_eq!(v["storage"]["files"][0]["mode"], 420);
-        assert_eq!(v["storage"]["files"][0]["overwrite"], true);
-        assert_eq!(v["storage"]["files"][0]["contents"]["source"], "data:,myvm");
-        assert!(v.get("passwd").is_none());
-    }
-
-    #[test]
-    fn test_storage_none_omitted_from_json() {
-        let cfg = IgnitionConfig {
-            ignition: IgnitionMeta {
-                version: "3.0.0".into(),
-            },
-            storage: None,
-            passwd: None,
-        };
-
-        let v: serde_json::Value =
-            serde_json::from_str(&serde_json::to_string_pretty(&cfg).unwrap()).unwrap();
-        assert!(v.get("storage").is_none());
-        assert!(v.get("passwd").is_none());
     }
 }
